+++ behave/configuration.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from __future__ import print_function
 import os
 import re
 import sys
@@ -240,19 +241,6 @@ options = [
      dict(action='append', dest='outfiles', metavar='FILE',
           help="Write to specified file instead of stdout.")),
 
-    (('--steps-prefix',),
-     dict(help="""Prepend the given prefix to the name of the steps directory
-                  and environment.py. Example: --steps-prefix=other_ will
-                  look for steps in the other_steps directory for each feature,
-                  and instead of environment.py, other_environment.py will be
-                  used.""",
-          config_help="""Prepend the given prefix to the name of the steps
-                         directory and environment.py. Example:
-                         ``steps-prefix = other_`` will look for steps in the
-                         other_steps directory for each feature, and instead of
-                         environment.py, other_environment.py will be
-                         used.""")),
-
     ((),  # -- CONFIGFILE only
      dict(action='append', dest='paths',
           help="Specify default feature paths, used when none are provided.")),
@@ -272,6 +260,12 @@ options = [
                   definition with the steps. This is the default
                   behaviour. This switch is used to override a
                   configuration file setting.""")),
+
+    (('--stage',),
+     dict(help="""Defines the current test stage.
+                  The test stage name is used as name prefix for the environment
+                  file and the steps directory (instead of default path names).
+                  """)),
 
     (('--stop',),
      dict(action='store_true',
@@ -386,7 +380,8 @@ def read_configuration(path):
                 outfiles.append(outfile)
             result['outfiles'] = outfiles
         elif len(outfiles) > formatter_size:
-            print "CONFIG-ERROR: Too many outfiles (%d) provided." % outfiles_size
+            print("CONFIG-ERROR: Too many outfiles (%d) provided." %
+                  outfiles_size)
             result['outfiles'] = outfiles[:formatter_size]
 
     for paths_name in ('paths', 'outfiles'):
@@ -410,13 +405,13 @@ def load_configuration(defaults, verbose=False):
             filename = os.path.join(path, filename)
             if os.path.isfile(filename):
                 if verbose:
-                    print 'Loading config defaults from "%s"' % filename
+                    print('Loading config defaults from "%s"' % filename)
                 defaults.update(read_configuration(filename))
 
     if verbose:
-        print 'Using defaults:'
+        print('Using defaults:')
         for k, v in defaults.items():
-            print '%15s %s' % (k, v)
+            print('%15s %s' % (k, v))
 
 
 # construct the parser
@@ -458,9 +453,7 @@ class Configuration(object):
         logging_level=logging.INFO,
         summary=True,
         junit=False,
-        steps_prefix='',
-        steps_dir='steps',
-        env_py='environment.py',
+        stage=None,
         # -- SPECIAL:
         default_format="pretty",   # -- Used when no formatters are configured.
         scenario_outline_annotation_schema=u"{name} -- @{row.id} {examples.name}"
@@ -502,6 +495,8 @@ class Configuration(object):
         self.include_re = None
         self.exclude_re = None
         self.scenario_outline_annotation_schema = None
+        self.steps_dir = "steps"
+        self.environment_file = "environment.py"
         if load_config:
             load_configuration(self.defaults, verbose=verbose)
         parser.set_defaults(**self.defaults)
@@ -562,9 +557,10 @@ class Configuration(object):
         if unknown_formats:
             parser.error("format=%s is unknown" % ", ".join(unknown_formats))
 
-        self.steps_dir = self.steps_prefix + 'steps'
-        self.env_py = self.steps_prefix + 'environment.py'
-
+        if self.stage is None:
+            # -- USE ENVIRONMENT-VARIABLE, if stage is undefined.
+            self.stage = os.environ.get("BEHAVE_STAGE", None)
+        self.setup_stage(self.stage)
         self.setup_model()
 
     def collect_unknown_formats(self):
@@ -637,3 +633,32 @@ class Configuration(object):
         if self.scenario_outline_annotation_schema:
             name_schema = unicode(self.scenario_outline_annotation_schema)
             ScenarioOutline.annotation_schema = name_schema.strip()
+
+    def setup_stage(self, stage=None):
+        """Setup the test stage that selects a different set of
+        steps and environment implementations.
+
+        :param stage:   Name of current test stage (as string or None).
+
+        EXAMPLE::
+
+            # -- SETUP DEFAULT TEST STAGE (unnamed):
+            config = Configuration()
+            config.setup_stage()
+            assert config.steps_dir == "steps"
+            assert config.environment_file == "environment.py"
+
+            # -- SETUP PRODUCT TEST STAGE:
+            config.setup_stage("product")
+            assert config.steps_dir == "product_steps"
+            assert config.environment_file == "product_environment.py"
+        """
+        steps_dir = "steps"
+        environment_file = "environment.py"
+        if stage:
+            # -- USE A TEST STAGE: Select different set of implementations.
+            prefix = stage + "_"
+            steps_dir = prefix + steps_dir
+            environment_file = prefix + environment_file
+        self.steps_dir = steps_dir
+        self.environment_file = environment_file