+++ behave/log_capture.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from __future__ import print_function
 import logging
 import functools
 from logging.handlers import BufferingHandler
@@ -222,8 +225,8 @@ def capture(*args, **kw):
                 h.abandon()
             v = h.getvalue()
             if v:
-                print 'Captured Logging:'
-                print v
+                print('Captured Logging:')
+                print(v)
         return f
 
     if not args: