+++ behave/runner.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from __future__ import with_statement
+from __future__ import with_statement, print_function
 import contextlib
 import os.path
 import StringIO
@@ -200,8 +200,8 @@ class Context(object):
 
     def _dump(self):
         for level, frame in enumerate(self._stack):
-            print 'Level %d' % level
-            print repr(frame)
+            print('Level %d' % level)
+            print(repr(frame))
 
     def __getattr__(self, attr):
         if attr[0] == '_':
@@ -496,14 +496,14 @@ class ModelRunner(object):
 
         # -- AFTER-ALL:
         if self.aborted:
-            print "\nABORTED: By user."
+            print("\nABORTED: By user.")
         for formatter in self.formatters:
             formatter.close()
         self.run_hook('after_all', self.context)
         for reporter in self.config.reporters:
             reporter.end()
         # if self.aborted:
-        #     print "\nABORTED: By user."
+        #     print("\nABORTED: By user.")
         failed = ((failed_count > 0) or self.aborted or
                   (len(self.undefined_steps) > undefined_steps_initial_size))
         return failed
@@ -534,8 +534,8 @@ class Runner(ModelRunner):
     def setup_paths(self):
         if self.config.paths:
             if self.config.verbose:
-                print 'Supplied path:', \
-                      ', '.join('"%s"' % path for path in self.config.paths)
+                print('Supplied path:', \
+                      ', '.join('"%s"' % path for path in self.config.paths))
             first_path = self.config.paths[0]
             if hasattr(first_path, "filename"):
                 # -- BETTER: isinstance(first_path, FileLocation):
@@ -552,24 +552,26 @@ class Runner(ModelRunner):
             # supplied path might be to a feature file
             if os.path.isfile(base_dir):
                 if self.config.verbose:
-                    print 'Primary path is to a file so using its directory'
+                    print('Primary path is to a file so using its directory')
                 base_dir = os.path.dirname(base_dir)
         else:
             if self.config.verbose:
-                print 'Using default path "./features"'
+                print('Using default path "./features"')
             base_dir = os.path.abspath('features')
 
         # Get the root. This is not guaranteed to be '/' because Windows.
         root_dir = path_getrootdir(base_dir)
         new_base_dir = base_dir
+        steps_dir = self.config.steps_dir
+        environment_file = self.config.environment_file
 
         while True:
             if self.config.verbose:
-                print 'Trying base directory:', new_base_dir
+                print('Trying base directory:', new_base_dir)
 
-            if os.path.isdir(os.path.join(new_base_dir, self.config.steps_dir)):
+            if os.path.isdir(os.path.join(new_base_dir, steps_dir)):
                 break
-            if os.path.isfile(os.path.join(new_base_dir, self.config.env_py)):
+            if os.path.isfile(os.path.join(new_base_dir, environment_file)):
                 break
             if new_base_dir == root_dir:
                 break
@@ -579,16 +581,14 @@ class Runner(ModelRunner):
         if new_base_dir == root_dir:
             if self.config.verbose:
                 if not self.config.paths:
-                    print 'ERROR: Could not find "%s" directory. Please '\
-                        'specify where to find your features.' \
-                            % self.config.steps_dir
+                    print('ERROR: Could not find "%s" directory. '\
+                          'Please specify where to find your features.' % \
+                                steps_dir)
                 else:
-                    print 'ERROR: Could not find "%s" directory in your '\
-                        'specified path "%s"' \
-                            % (self.config.steps_dir, base_dir)
+                    print('ERROR: Could not find "%s" directory in your '\
+                        'specified path "%s"' % (steps_dir, base_dir))
 
-            message = 'No %s directory in "%s"' \
-                          % (self.config.steps_dir, base_dir)
+            message = 'No %s directory in "%s"' % (steps_dir, base_dir)
             raise ConfigError(message)
 
         base_dir = new_base_dir
@@ -600,11 +600,11 @@ class Runner(ModelRunner):
         else:
             if self.config.verbose:
                 if not self.config.paths:
-                    print 'ERROR: Could not find any "<name>.feature" files. '\
-                        'Please specify where to find your features.'
+                    print('ERROR: Could not find any "<name>.feature" files. '\
+                        'Please specify where to find your features.')
                 else:
-                    print 'ERROR: Could not find any "<name>.feature" files '\
-                        'in your specified path "%s"' % base_dir
+                    print('ERROR: Could not find any "<name>.feature" files '\
+                        'in your specified path "%s"' % base_dir)
             raise ConfigError('No feature files in "%s"' % base_dir)
 
         self.base_dir = base_dir
@@ -622,8 +622,8 @@ class Runner(ModelRunner):
         """
         context.config.setup_logging()
 
-    def load_hooks(self, filename=''):
-        filename = filename or self.config.env_py
+    def load_hooks(self, filename=None):
+        filename = filename or self.config.environment_file
         hooks_path = os.path.join(self.base_dir, filename)
         if os.path.exists(hooks_path):
             exec_file(hooks_path, self.hooks)